+++ src/Rest/ResourceActionGrantProvider.php
@@ -6,7 +6,6 @@ namespace Dbp\Relay\AuthorizationBundle\Rest;
 
 use Dbp\Relay\AuthorizationBundle\Authorization\AuthorizationService;
 use Dbp\Relay\AuthorizationBundle\Entity\ResourceActionGrant;
-use Dbp\Relay\AuthorizationBundle\Service\InternalResourceActionGrantService;
 use Dbp\Relay\CoreBundle\Rest\AbstractDataProvider;
 use Dbp\Relay\CoreBundle\Rest\Query\Pagination\Pagination;
 
@@ -18,7 +17,6 @@ use Dbp\Relay\CoreBundle\Rest\Query\Pagination\Pagination;
 class ResourceActionGrantProvider extends AbstractDataProvider
 {
     public function __construct(
-        private readonly InternalResourceActionGrantService $resourceActionGrantService,
         private readonly AuthorizationService $authorizationService)
     {
         parent::__construct();
@@ -26,7 +24,7 @@ class ResourceActionGrantProvider extends AbstractDataProvider
 
     protected function getItemById(string $id, array $filters = [], array $options = []): ?object
     {
-        return $this->resourceActionGrantService->getResourceActionGrantByIdentifier($id);
+        return $this->authorizationService->getResourceActionGrantByIdentifier($id);
     }
 
     protected function getPage(int $currentPageNumber, int $maxNumItemsPerPage, array $filters = [], array $options = []): array